requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
dateparser==1.2.1
Flask==2.3.3
FlightRadarAPI==1.3.34
frozenlist==1.7.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==5.7.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.5.0
oauth2client==4.1.3
oauthlib==3.3.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
pyparsing==3.2.3
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
